feat: product images 업로드 기능 추가 - AWS S3 사용
1. 이미지 파일 업로드 요청 (클라이언트 → 서버)
2. 서버가 S3에 업로드하고, URL을 응답으로 반환
3. 클라이언트가 이 URL들을 포함하여 상품 등록/수정 요청
4. 서버는 imageUrl을 ProductImage로 매핑하고, Product에 연관 저장
5. 상품 조회 시 imageUrl이 포함된 데이터 반환

+++ src/main/java/com/travelservice/domain/product/service/ProductServiceImpl.java
@@ -1,5 +1,6 @@
 package com.travelservice.domain.product.service;
 
+import java.io.IOException;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -21,11 +22,12 @@ public class ProductServiceImpl implements ProductService {
 
 	private final ProductRepository productRepository;
 	private final RegionRepository regionRepository;
+	private final S3Uploader s3Uploader;
 
 	@Override
-	public Product createProduct(AddProductRequest request) {
+	public Product createProduct(AddProductRequest request) throws IOException {
 		Region region = regionRepository.findById(request.getRegionId())
-			.orElseThrow(() -> new IllegalArgumentException("not found region: " + request.getRegionId()));
+			.orElseThrow(() -> new IllegalArgumentException("지역을 찾을 수 없습니다" + request.getRegionId()));
 
 		Product product = request.toEntity(region);
 		return productRepository.save(product);
@@ -39,7 +41,7 @@ public class ProductServiceImpl implements ProductService {
 	@Override
 	public Product getProductDetail(Integer productId) {
 		Product product = productRepository.findById(productId)
-			.orElseThrow(() -> new EntityNotFoundException("해당 상품을 찾을 수 없습니다"));
+			.orElseThrow(() -> new EntityNotFoundException("상품을 찾을 수 없습니다: " + productId));
 
 		return product;
 	}
@@ -48,13 +50,11 @@ public class ProductServiceImpl implements ProductService {
 	@Transactional
 	public Product updateProduct(Integer productId, UpdateProductRequest request) {
 		Product product = productRepository.findById(productId)
-			.orElseThrow(() -> new IllegalArgumentException("not found: " + productId));
+			.orElseThrow(() -> new IllegalArgumentException("상품을 찾을 수 없습니다: " + productId));
 
 		Region region = regionRepository.findById(request.getRegionId())
-			.orElseThrow(() -> new RuntimeException("not found: region " + request.getRegionId()));
+			.orElseThrow(() -> new RuntimeException("지역을 찾을 수 없습니다 " + request.getRegionId()));
 
-		// product.update(request.getName(), request.getPrice(), request.getTotalQuantity(), request.getStockQuantity(),
-		// 	request.getDescription(), request.getSaleStatus(), request.getType(), request.getDuration(), region);
 		request.applyTo(product, region);
 
 		return product;
@@ -63,8 +63,9 @@ public class ProductServiceImpl implements ProductService {
 	@Override
 	@Transactional
 	public void deleteProduct(Integer productId) {
-		Product product = productRepository.findById(productId)
-			.orElseThrow(() -> new IllegalArgumentException("존재하지 않는 상품입니다"));
+		if (!productRepository.existsById(productId)) {
+			throw new IllegalArgumentException("존재하지 않는 상품입니다");
+		}
 		productRepository.deleteById(productId);
 	}
 }
